// src/types/Reward.ts
type Reward = {
  karma_gained: number;
  created_vents_counter: number;
  created_vent_supports_counter: number;
  received_vent_supports_counter: number;
  created_comments_counter: number;
  created_comment_supports_counter: number;
  received_comment_supports_counter: number;
  created_quotes_counter: number;
  created_quote_supports_counter: number;
  received_quote_supports_counter: number;
  quote_contests_won_counter: number;
  server_timestamp: number;
  title: string;
};

export default Reward;